% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeparg_convert2otu.R
\name{deeparg_convert2otu}
\alias{deeparg_convert2otu}
\title{DeepARG}
\arguments{
\item{input}{Something}
}
\description{
This function allows you to convert DeepARG into Ampvis2 compatible format
}
\details{
A function for converting DeepARG outputs

something
}
\examples{
deeparg_convert2otu(deeparg_input)
}
\keyword{Ampvis2}
\keyword{DeepARG}
