% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graftm_convert2otu.R
\name{graftm_convert2otu}
\alias{graftm_convert2otu}
\title{GraftM}
\arguments{
\item{input}{Somethin}
}
\description{
This function allows you to convert GraftM into Ampvis2 compatible format
}
\details{
A function for converting GraftM outputs

something
}
\examples{
graftm_convert2otu(graftm_input)
}
\keyword{Ampvis2}
\keyword{GraftM}
