% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaiju_convert2otu.R
\name{kaiju_convert2otu}
\alias{kaiju_convert2otu}
\title{Kaiju}
\arguments{
\item{input}{Something}
}
\description{
This function allows you to convert Kaiju into Ampvis2 compatible format
}
\details{
A function for converting Kaiju outputs

something
}
\examples{
kaiju_convert2otu(kaiju_input)
}
\keyword{Ampvis2}
\keyword{GraftM}
