% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken2_convert2otu.R
\name{kraken2_convert2otu}
\alias{kraken2_convert2otu}
\title{Kraken2}
\arguments{
\item{input}{Something}
}
\description{
This function allows you to convert Kraken2 into Ampvis2 compatible format
}
\details{
A function for converting Kraken2 outputs

something
}
\examples{
kraken2_convert2otu(kraken2_input)
}
\keyword{Ampvis2}
\keyword{Kraken2}
