% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimap2card_convert2otu.R
\name{minimap2card_convert2otu}
\alias{minimap2card_convert2otu}
\title{minimap2card}
\arguments{
\item{input}{Somethin}
}
\description{
This function allows you to convert minimap2card into Ampvis2 compatible format
}
\details{
A function for converting minimap2card outputs

something
}
\examples{
minimap2card_convert2otu(minimap2card_input)
}
\keyword{Ampvis2}
\keyword{CARD}
\keyword{minimap2}
