% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlem_convert2otu.R
\name{singlem_convert2otu}
\alias{singlem_convert2otu}
\title{SingleM}
\arguments{
\item{input}{Something}
}
\description{
This function allows you to convert SingleM into Ampvis2 compatible format
}
\details{
A function for converting SingleM outputs

something
}
\examples{
singlem_convert2otu(singlem_input)
}
\keyword{Ampvis2}
\keyword{SingleM}
